move avs & operator rewards routes to their own controllers
