+++ packages/api/src/routes/operators/operatorController.ts
@@ -1,18 +1,30 @@
-import type Prisma from '@prisma/client'
 import type { Request, Response } from 'express'
-import prisma from '../../utils/prismaClient'
+import {
+	type EigenStrategiesContractAddress,
+	getEigenContracts
+} from '../../data/address'
 import { PaginationQuerySchema } from '../../schema/zod/schemas/paginationQuery'
 import { EthereumAddressSchema } from '../../schema/zod/schemas/base/ethereumAddress'
 import { WithTvlQuerySchema } from '../../schema/zod/schemas/withTvlQuery'
-import { handleAndReturnErrorResponse } from '../../schema/errors'
-import { fetchStrategyTokenPrices } from '../../utils/tokenPrices'
-import {
-	getStrategiesWithShareUnderlying,
-	sharesToTVL
-} from '../strategies/strategiesController'
 import { WithAdditionalDataQuerySchema } from '../../schema/zod/schemas/withAdditionalDataQuery'
 import { SortByQuerySchema } from '../../schema/zod/schemas/sortByQuery'
 import { SearchByTextQuerySchema } from '../../schema/zod/schemas/searchByTextQuery'
+import { RewardsQuerySchema } from '../../schema/zod/schemas/rewardsQuery'
+import { handleAndReturnErrorResponse } from '../../schema/errors'
+import {
+	fetchRewardTokenPrices,
+	fetchStrategyTokenPrices
+} from '../../utils/tokenPrices'
+import {
+	getStrategiesWithShareUnderlying,
+	sharesToTVL,
+	sharesToTVLEth
+} from '../strategies/strategiesController'
+import { withOperatorShares } from '../avs/avsController'
+import { getNetwork } from '../../viem/viemClient'
+import { holesky } from 'viem/chains'
+import Prisma from '@prisma/client'
+import prisma from '../../utils/prismaClient'
 
 /**
  * Function for route /operators
@@ -131,9 +143,9 @@ export async function getAllOperators(req: Request, res: Response) {
  */
 export async function getOperator(req: Request, res: Response) {
 	// Validate pagination query
-	const result = WithTvlQuerySchema.and(
-		WithAdditionalDataQuerySchema
-	).safeParse(req.query)
+	const result = WithTvlQuerySchema.and(WithAdditionalDataQuerySchema)
+		.and(RewardsQuerySchema)
+		.safeParse(req.query)
 	if (!result.success) {
 		return handleAndReturnErrorResponse(req, res, result.error)
 	}
@@ -143,7 +155,7 @@ export async function getOperator(req: Request, res: Response) {
 		return handleAndReturnErrorResponse(req, res, paramCheck.error)
 	}
 
-	const { withTvl, withAvsData } = result.data
+	const { withTvl, withAvsData, apy } = result.data
 
 	try {
 		const { address } = req.params
@@ -155,27 +167,48 @@ export async function getOperator(req: Request, res: Response) {
 					select: {
 						avsAddress: true,
 						isActive: true,
-						...(withAvsData
+						...(withAvsData || apy
 							? {
 									avs: {
 										select: {
-											metadataUrl: true,
-											metadataName: true,
-											metadataDescription: true,
-											metadataDiscord: true,
-											metadataLogo: true,
-											metadataTelegram: true,
-											metadataWebsite: true,
-											metadataX: true,
-											curatedMetadata: true,
-											restakeableStrategies: true,
-											totalStakers: true,
-											totalOperators: true,
-											tvlEth: true,
-											createdAtBlock: true,
-											updatedAtBlock: true,
-											createdAt: true,
-											updatedAt: true
+											...(withAvsData
+												? {
+														metadataUrl: true,
+														metadataName: true,
+														metadataDescription: true,
+														metadataDiscord: true,
+														metadataLogo: true,
+														metadataTelegram: true,
+														metadataWebsite: true,
+														metadataX: true,
+														curatedMetadata: true,
+														restakeableStrategies: true,
+														totalStakers: true,
+														totalOperators: true,
+														tvlEth: true,
+														createdAtBlock: true,
+														updatedAtBlock: true,
+														createdAt: true,
+														updatedAt: true
+												  }
+												: {}),
+											...(apy
+												? {
+														address: true,
+														rewardSubmissions: true,
+														restakeableStrategies: true,
+														operators: {
+															where: { isActive: true },
+															include: {
+																operator: {
+																	include: {
+																		shares: true
+																	}
+																}
+															}
+														}
+												  }
+												: {})
 										}
 									}
 							  }
@@ -209,6 +242,7 @@ export async function getOperator(req: Request, res: Response) {
 						strategyTokenPrices
 				  )
 				: undefined,
+			rewards: apy ? await calculateOperatorApy(operator) : undefined,
 			stakers: undefined,
 			metadataUrl: undefined,
 			isMetadataSynced: undefined,
@@ -222,7 +256,7 @@ export async function getOperator(req: Request, res: Response) {
 }
 
 /**
- * Function for route /operator/addresses
+ * Function for route /operators/addresses
  * Returns a list of all Operators, addresses & logos. Optionally perform a text search for a list of matched Operators.
  *
  * @param req
@@ -292,6 +326,89 @@ export async function getAllOperatorAddresses(req: Request, res: Response) {
 }
 
 /**
+ * Function for route /operators/:address/rewards
+ * Returns a list of strategies that the Operator is rewarded for, and the tokens they're rewarded in
+ *
+ * @param req
+ * @param res
+ * @returns
+ */
+export async function getOperatorRewards(req: Request, res: Response) {
+	const paramCheck = EthereumAddressSchema.safeParse(req.params.address)
+	if (!paramCheck.success) {
+		return handleAndReturnErrorResponse(req, res, paramCheck.error)
+	}
+
+	try {
+		const { address } = req.params
+
+		// Fetch Operator data
+		const operator = await prisma.operator.findUnique({
+			where: { address: address.toLowerCase() },
+			include: {
+				avs: {
+					include: {
+						avs: {
+							include: {
+								rewardSubmissions: true
+							}
+						}
+					}
+				},
+				shares: true
+			}
+		})
+
+		if (!operator) {
+			throw new Error('Operator not found.')
+		}
+
+		const result: {
+			operatorAddress: string
+			rewardTokens: Set<string>
+			strategiesRewarded: Set<string>
+		} = {
+			operatorAddress: address,
+			rewardTokens: new Set<string>(),
+			strategiesRewarded: new Set<string>()
+		}
+
+		// Create a Set of strategies where the operator has positive TVL
+		const operatorActiveStrategies = new Set(
+			operator.shares
+				.filter((share) => new Prisma.Prisma.Decimal(share.shares).gt(0))
+				.map((share) => share.strategyAddress.toLowerCase())
+		)
+
+		// Iterate through all Avs
+		for (const avsOperator of operator.avs) {
+			const avs = avsOperator.avs
+
+			// Iterate through all reward submissions
+			for (const submission of avs.rewardSubmissions) {
+				result.rewardTokens.add(submission.token.toLowerCase())
+
+				if (
+					operatorActiveStrategies.has(submission.strategyAddress.toLowerCase())
+				) {
+					result.strategiesRewarded.add(
+						submission.strategyAddress.toLowerCase()
+					)
+				}
+			}
+		}
+
+		res.send({
+			operatorAddress: result.operatorAddress,
+			rewardTokens: Array.from(result.rewardTokens),
+			strategiesRewarded: Array.from(result.strategiesRewarded)
+		})
+	} catch (error) {
+		handleAndReturnErrorResponse(req, res, error)
+	}
+}
+
+/**
  * Function for route /operators/:address/invalidate-metadata
  * Protected route to invalidate the metadata of a given Operator
  *
@@ -322,6 +439,8 @@ export async function invalidateMetadata(req: Request, res: Response) {
 	}
 }
 
+// --- Helper functions ---
+
 export function getOperatorSearchQuery(
 	searchByText: string | undefined,
 	searchMode: 'contains' | 'startsWith',
@@ -348,4 +467,221 @@ export function getOperatorSearchQuery(
 			{ metadataWebsite: searchConfig }
 		] as Prisma.Prisma.OperatorWhereInput[]
 	}
+}
+
+// biome-ignore lint/suspicious/noExplicitAny: <explanation>
+async function calculateOperatorApy(operator: any) {
+	try {
+		const avsRewardsMap: Map<string, number> = new Map()
+		const strategyRewardsMap: Map<string, number> = new Map()
+
+		// Grab the all reward submissions that the Operator is eligible for basis opted strategies & AVSs
+		const optedStrategyAddresses: Set<string> = new Set(
+			operator?.shares.map((share) => share.strategyAddress.toLowerCase())
+		)
+		const avsWithEligibleRewardSubmissions = operator?.avs
+			.filter((avsOp) => avsOp.avs.rewardSubmissions.length > 0)
+			.map((avsOp) => ({
+				avs: avsOp.avs,
+				eligibleRewards: avsOp.avs.rewardSubmissions.filter((reward) =>
+					optedStrategyAddresses.has(reward.strategyAddress.toLowerCase())
+				)
+			}))
+			.filter((item) => item.eligibleRewards.length > 0)
+
+		if (!avsWithEligibleRewardSubmissions) {
+			return {
+				avs: [],
+				strategies: [],
+				aggregateApy: 0,
+				averageAvsApy: 0,
+				averageStrategiesApy: 0,
+				operatorEarningsEth: 0
+			}
+		}
+
+		let operatorEarningsEth = new Prisma.Prisma.Decimal(0)
+
+		const strategyTokenPrices = await fetchStrategyTokenPrices()
+		const rewardTokenPrices = await fetchRewardTokenPrices()
+		const eigenContracts = getEigenContracts()
+		const tokenToStrategyMap = tokenToStrategyAddressMap(
+			eigenContracts.Strategies
+		)
+
+		const strategiesWithSharesUnderlying =
+			await getStrategiesWithShareUnderlying()
+
+		// Calc aggregate APY for each AVS basis the opted-in strategies
+		for (const avs of avsWithEligibleRewardSubmissions) {
+			let aggregateApy = 0
+
+			// Get share amounts for each restakeable strategy
+			const shares = withOperatorShares(avs.avs.operators).filter(
+				(s) =>
+					avs.avs.restakeableStrategies.indexOf(
+						s.strategyAddress.toLowerCase()
+					) !== -1
+			)
+
+			// Fetch the AVS tvl for each strategy
+			const tvlStrategiesEth = sharesToTVLEth(
+				shares,
+				strategiesWithSharesUnderlying,
+				strategyTokenPrices
+			)
+
+			// Iterate through each strategy and calculate all its rewards
+			for (const strategyAddress of optedStrategyAddresses) {
+				const strategyTvl = tvlStrategiesEth[strategyAddress.toLowerCase()] || 0
+				if (strategyTvl === 0) continue
+
+				let totalRewardsEth = new Prisma.Prisma.Decimal(0)
+				let totalDuration = 0
+
+				// Find all reward submissions attributable to the strategy
+				const relevantSubmissions = avs.eligibleRewards.filter(
+					(submission) =>
+						submission.strategyAddress.toLowerCase() ===
+						strategyAddress.toLowerCase()
+				)
+
+				for (const submission of relevantSubmissions) {
+					let rewardIncrementEth = new Prisma.Prisma.Decimal(0)
+					const rewardTokenAddress = submission.token.toLowerCase()
+					const tokenStrategyAddress =
+						tokenToStrategyMap.get(rewardTokenAddress)
+
+					// Normalize reward amount to its ETH price
+					if (tokenStrategyAddress) {
+						const tokenPrice = Object.values(strategyTokenPrices).find(
+							(tp) => tp.strategyAddress.toLowerCase() === tokenStrategyAddress
+						)
+						rewardIncrementEth = submission.amount.mul(
+							new Prisma.Prisma.Decimal(tokenPrice?.eth ?? 0)
+						)
+					} else {
+						// Check if it is a reward token which isn't a strategy on EL
+						for (const [, price] of Object.entries(rewardTokenPrices)) {
+							if (
+								price &&
+								price.tokenAddress.toLowerCase() === rewardTokenAddress
+							) {
+								rewardIncrementEth = submission.amount.mul(
+									new Prisma.Prisma.Decimal(price.eth ?? 0)
+								)
+							} else {
+								// Check for special tokens
+								rewardIncrementEth = isSpecialToken(rewardTokenAddress)
+									? submission.amount
+									: new Prisma.Prisma.Decimal(0)
+							}
+						}
+					}
+
+					// Multiply reward amount in ETH by the strategy weight
+					rewardIncrementEth = rewardIncrementEth
+						.mul(submission.multiplier)
+						.div(new Prisma.Prisma.Decimal(10).pow(18))
+
+					// Operator takes 10% in commission
+					const operatorFeesEth = rewardIncrementEth.mul(10).div(100)
+					operatorEarningsEth = operatorEarningsEth.add(operatorFeesEth)
+
+					totalRewardsEth = totalRewardsEth
+						.add(rewardIncrementEth)
+						.sub(operatorFeesEth)
+					totalDuration += submission.duration
+				}
+
+				if (totalDuration === 0) continue
+
+				// Annualize the reward basis its duration to find yearly APY
+				const rewardRate =
+					totalRewardsEth
+						.div(new Prisma.Prisma.Decimal(10).pow(18))
+						.toNumber() / strategyTvl
+				const annualizedRate =
+					rewardRate * ((365 * 24 * 60 * 60) / totalDuration)
+				const apy = annualizedRate * 100
+				aggregateApy += apy
+
+				// Add strategy's APY to common strategy rewards store (across all Avs)
+				const currentStrategyApy = strategyRewardsMap.get(strategyAddress) || 0
+				strategyRewardsMap.set(strategyAddress, currentStrategyApy + apy)
+			}
+			// Add aggregate APY to Avs rewards store
+			avsRewardsMap.set(avs.avs.address, aggregateApy)
+		}
+
+		const response = {
+			avs: Array.from(avsRewardsMap, ([avsAddress, apy]) => ({
+				avsAddress,
+				apy
+			})),
+			strategies: Array.from(strategyRewardsMap, ([strategyAddress, apy]) => ({
+				strategyAddress,
+				apy
+			})),
+			aggregateApy: 0,
+			averageAvsApy: 0,
+			averageStrategiesApy: 0,
+			operatorEarningsEth: new Prisma.Prisma.Decimal(0)
+		}
+
+		// Calculate aggregate and averages across Avs and strategies
+		response.aggregateApy = response.avs.reduce((sum, avs) => sum + avs.apy, 0)
+		response.averageAvsApy =
+			response.avs.length > 0 ? response.aggregateApy / response.avs.length : 0
+		response.averageStrategiesApy =
+			response.strategies.length > 0
+				? response.strategies.reduce((sum, strategy) => sum + strategy.apy, 0) /
+				  response.strategies.length
+				: 0
+		response.operatorEarningsEth = operatorEarningsEth
+
+		return response
+	} catch {}
+}
+
+/**
+ * Return a map of strategy addresses <> token addresses
+ *
+ * @param strategies
+ * @returns
+ */
+export function tokenToStrategyAddressMap(
+	strategies: EigenStrategiesContractAddress
+): Map<string, string> {
+	const map = new Map<string, string>()
+	for (const [key, value] of Object.entries(strategies)) {
+		if (key !== 'Eigen' && value?.tokenContract && value?.strategyContract) {
+			map.set(
+				value.tokenContract.toLowerCase(),
+				value.strategyContract.toLowerCase()
+			)
+		}
+	}
+	return map
+}
+
+/**
+ * Returns whether a given token address belongs to a list of special tokens
+ *
+ * @param tokenAddress
+ * @returns
+ */
+export function isSpecialToken(tokenAddress: string): boolean {
+	const specialTokens =
+		getNetwork() === holesky
+			? [
+					'0x6Cc9397c3B38739daCbfaA68EaD5F5D77Ba5F455', // WETH
+					'0xbeac0eeeeeeeeeeeeeeeeeeeeeeeeeeeeeebeac0'
+			  ]
+			: [
+					'0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2', // WETH
+					'0xbeac0eeeeeeeeeeeeeeeeeeeeeeeeeeeeeebeac0'
+			  ]
+
+	return specialTokens.includes(tokenAddress.toLowerCase())
 }
