+++ packages/api/src/routes/operators/operatorRoutes.ts
@@ -3,6 +3,7 @@ import {
 	getAllOperators,
 	getOperator,
 	getAllOperatorAddresses,
+	getOperatorRewards,
 	invalidateMetadata
 } from './operatorController'
 import { authenticateJWT } from '../../utils/jwtUtils'
@@ -104,6 +105,12 @@ router.get('/', routeCache.cacheSeconds(120), getAllOperators)
  *               example: 'Operator not found.'
  */
 router.get('/:address', routeCache.cacheSeconds(120), getOperator)
+
+router.get(
+	'/:address/rewards',
+	routeCache.cacheSeconds(120),
+	getOperatorRewards
+)
 
 // Protected routes
 router.get(
